Fix load scenario
